requirements.txt
aiofiles==23.2.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
asyncpg==0.28.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
clipboard==0.0.4
colorlog==6.7.0
fastapi==0.101.1
fastjsonschema==2.18.0
Flask==2.1.3
Flask-Cors==4.0.0
flask-sock==0.6.0
h11==0.14.0
html5tagger==1.3.0
httptools==0.6.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
lxml==4.9.3
MarkupSafe==2.0.1
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pydantic==2.1.1
pydantic_core==2.4.0
pyperclip==1.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
reactpy==1.0.2
reactpy-router==0.1.0
requests==2.31.0
sanic==23.6.0
Sanic-Cors==2.2.0
sanic-routing==23.6.0
simple-websocket==0.10.1
sniffio==1.3.0
starlette==0.27.0
tornado==6.3.3
tracerite==1.1.0
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.1.2
wsproto==1.2.0
